Add example responses for GET /ob/profile and GET /ob/listing
The profile example isn't a complete example of all of the possible fields that can be added; more work is required to get that example and infer the schema.

The listing example is for a physical listing with options/variants. In future we should test with dedicated schema for digital and service listings.